pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-yoink"
version = "1.0.0"
description = "Complete Discord server backup and recreation tool"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Magnus Naasade", email = "magn6162@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: System :: Archiving :: Backup"
]
keywords = ["discord", "backup", "server", "chat", "archive"]
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.2",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pillow>=9.0.0",
    "python-dateutil>=2.8.0",
    "tqdm>=4.64.0",
    "click>=8.1.0",
    "jinja2>=3.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/mnaljm/Project-yoink"
Repository = "https://github.com/mnaljm/Project-yoink.git"
Issues = "https://github.com/mnaljm/Project-yoink/issues"

[project.scripts]
discord-yoink = "discord_yoink:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
